Correct popwin configuration
When I updated to Emacs 26, a weird thing started happening: Magit
status buffers weren't full-screen! Actually, a lot of buffers were
appearing on the right half of the screen when they didn't do so in
Emacs 25.

At first I suspect the new Emacs 26 feature of "side windows", but I
found that side windows had entirely different behaviour from what I
was seeing. So that couldn't be it.

I dug around the various "action" functions you can provide to
`display-buffer`, and weirdly they all produced the same behaviour.

Eventually I started putting `edebug` on the Magit buffer management
functions, on `display-buffer` and related functions, and found a
conditional saying something like:

    (if display-buffer-function
        (funcall display-buffer-function ... )
      ... )

And I saw that `popwin` used as `display-buffer-function`.

Except Popwin is supposed to be enabled only on things which match
`popwin:special-display-config`, and it was activating everywhere.

So I put `edebug` on a bunch of `popwin` functions, and found
`popwin:original-display-buffer`, which is *supposed to* call the
unmodified `display-buffer` function. Except `display-buffer` was
checking `display-buffer-function`, sending us back to `popwin`!

So basically, setting `display-buffer-function` to `popwin` is 100%
the wrong way to do things. Now the config uses the minor mode as
intended, and things seem to work correctly.

I removed a couple seemingly useless things from the configuration, as
well. vc-log, some shell related stuff, some stuff for Slime, etc.